New Antagonist: The Spider Queen
Queen abilities:

— Laying eggs (4 pieces).
— Summons three small spiders that disappear after 20 seconds.
— Create a wall of web, the web itself and a blue crystal for lighting.

Spiders:

— Spider Queen (the strongest unit, although it is not difficult to kill, so it relies on its children and laid eggs).
— Spider Hunter (the main force of the hive, the best bite).
— Spider Guard (can build walls and webs with crystals).
— Ghost Spider (has invisibility, deals poison damage and injects a sleeping reagent).
— Spider Shooter (shoots poison, hits with poison)

You could say that it will be similar to the Rat King, only the servants will be players. They, as a rule, do not attack themselves, but simply create problems in technical rooms. You can kill one spider with a crowbar, their strength is in the web and unity!

Right now it's not 100% ready, at least the description is normal, the names and distribution into folders are necessary, but the gameplay is all there, so you can take a look and say what to change or something like that.
Antagonist on 100% prototypes